検索機能

+++ keruhito/app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Http\Requests\PostRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 
@@ -17,7 +18,7 @@ class PostController extends Controller
      */
     public function index()
     {
-        $posts = Post::latest()->paginate(5);
+        $posts = Post::latest()->paginate(10);
         return view('post.index', compact('posts'));
     }
 
@@ -82,6 +83,9 @@ class PostController extends Controller
     {
         $post = Post::findOrFail($id);
         $file_name = $post->image;
+        if ($post->user_id !== Auth::id()) {
+            return route('post.edit', ['id' => $post->id])->with('success', 'ユーザーを確認してください');
+        }
         if ($request->file('image')) {
             $file_name = mt_rand() . '.jpg';
             Storage::delete('public/images/' . $post->image);
@@ -104,5 +108,15 @@ class PostController extends Controller
     {
         Post::where('id', $id)->delete();
         return redirect()->route('post.index')->with('success', '削除しました');
+    }
+
+    public function search(Request $request)
+    {
+        $posts = Post::where('title', 'like', "%{$request->search}%")
+        ->orWhere('content', 'like', "%{$request->search}%")
+        ->paginate(10);
+
+        $search_result = '"' . $request->search . '"の検索結果：' . $posts->total() . '件';
+        return view('post.index', compact('posts', 'search_result'));
     }
 }
